+++ example/a.ts
@@ -1,11 +1,13 @@
-import { Elysia, t } from '../src'
+import { Elysia } from '../src'
 
-const plugin = async () => new Elysia().get('/', () => 'A')
+const a = (config = {}) =>
+	new Elysia({
+		name: 'a',
+		seed: config
+	}).get('/', () => 'a')
 
-const app = new Elysia()
-	.use(plugin())
-	.listen(3000)
+const app = new Elysia().use(a()).listen(3000)
+
+await app.modules
 
 console.log(app.routes)
-
-type App = typeof app