+++ test/core/modules.test.ts
@@ -124,4 +124,18 @@ describe('Modules', () => {
 
 		expect(res).toBe('hi')
 	})
+
+	describe('handle async plugin', async () => {
+		const a = (config = {}) =>
+			new Elysia({
+				name: 'a',
+				seed: config
+			}).get('/', () => 'a')
+
+		const app = new Elysia().use(a()).compile()
+
+		await app.modules
+
+		expect(app.routes.length).toBe(1)
+	})
 })